@@ -1,6 +1,6 @@
 import { Injectable } from '@angular/core';
 import { HttpClient, HttpParams } from "@angular/common/http";
-import { Observable } from "rxjs/Observable";
+import { Observable } from "rxjs";
 
 import { CreateTicket } from '../../entities/write/createTicket';
 import { WRITE_API } from "../../configuration/config";
