@@ -8,13 +8,8 @@ import { AddTicketDetails } from '../../redux/actions/ticketDetail';
 import { TicketDetail } from '../../entities/read/ticketDetail';
 import { DetailTicketService } from '../../containers/detail-ticket/detail-ticket.service';
 
-import { Observable } from 'rxjs/Observable';
-import { of } from 'rxjs/observable/of';
-import 'rxjs/add/operator/do';
-import 'rxjs/add/operator/take';
-import 'rxjs/add/operator/switchMap';
-import 'rxjs/add/operator/catch';
-import 'rxjs/add/operator/filter';
+import { Observable, pipe, of } from 'rxjs';
+import { take, switchMap, catchError, filter, tap } from 'rxjs/operators';
 
 @Injectable()
 export class TicketDetailGuard implements CanActivate {
@@ -27,17 +22,20 @@ export class TicketDetailGuard implements CanActivate {
         return this.store
             // selecting the ticket state using a feature selector
             .select(TicketDetailSelectors.getTicketDetail(id))
-            // the .do() operator allows for a side effect, at this
+            // the .tap() operator allows for a side effect, at this
             // point, I'm checking if the ticket property exists on my 
             // Store slice of state
-            .do((data: TicketDetail) => {
-                // if the ticket is not in the store add it trough the backend
-                if (!data) {
-                    this.service.get(id).take(1).subscribe(ticket => this.store.dispatch(new AddTicketDetails(ticket)));
-                }
-            })
-            .filter(data => data !== undefined)
-            .take(1);
+            .pipe(
+                tap((data: TicketDetail) => {
+                    // if the ticket is not in the store add it trough the backend
+                    if (!data) {
+                        this.service.get(id).pipe(take(1))
+                            .subscribe(ticket => this.store.dispatch(new AddTicketDetails(ticket)));
+                    }
+                }),
+                filter(data => data !== undefined),
+                take(1)
+            );
     }
 
     // our guard that gets called each time we 
@@ -46,9 +44,11 @@ export class TicketDetailGuard implements CanActivate {
         const id: string = route.params['id'];
         // return our Observable stream from above
         return this.getFromStoreOrAPI(id)
-            // if it was successful, we can return Observable.of(true)
-            .switchMap(() => of(true))
-            // otherwise, something went wrong
-            .catch(() => of(false));
+            .pipe(
+                // if it was successful, we can return Observable.of(true)
+                switchMap(() => of(true)),
+                // otherwise, something went wrong
+                catchError(() => of(false))
+            );
     }
 }
